fix: undo for prefab
