+++ Editor/Utility/FTPrefabUtility.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e6d915a0239a486459e9c6d7ed1a8c3d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
